changed type to double in pid

+++ navigator/navigator-system/src/main/java/ai/aitia/navigator/navigator_system/PIDController.java
@@ -2,20 +2,20 @@ package ai.aitia.navigator.navigator_system;
 
 public class PIDController {
 
-    private final float k;
-    private final float ki;
-    private final float kd;
+    private final double k;
+    private final double ki;
+    private final double kd;
 
-    private float last_e = 0;
-    private float integral_e = 0;
+    private double last_e = 0;
+    private double integral_e = 0;
 
-    public PIDController(float k, float ki, float kd) {
+    public PIDController(double k, double ki, double kd) {
         this.k = k;
         this.ki = ki;
         this.kd = kd;
     }
 
-    public float getNextU(float e) {
+    public double getNextU(double e) {
         integral_e += e;
         return (k * e + kd * (e - last_e) + kd * integral_e);
     }
